Added score tracking and fine-tuned restarting the game, added that too
The game can now be played to full 10 points, there is an on screen display that shows how many successes the player has achieved.  Once 10 are hit correctly, a restart button will appear, which will cycle the game back to the main menu layout, and begin the game anew.  All thats missing is a stopwatch.